Add and configure babel styled-components plugin
This adds the babel styled-components plugin, so that I can pass the
`css` props into components, instead of using the `styled` constructor.
The main advantage here is that this will work with jsx-a11y.

This uses babel macros, which require a special import. This also
requires a special type declaration, since the `css` prop is new to
JSX elements.